If localStorage is not available, use a in-memory none expiring assignment cache (FF-2110)
