+++ src/configuration-factory.ts
@@ -48,16 +48,3 @@ export function configurationStorageFactory(
   // No persistence store available, use memory only
   return new MemoryOnlyConfigurationStore();
 }
-
-export function hasChromeStorage(): boolean {
-  return typeof chrome !== 'undefined' && !!chrome.storage;
-}
-
-export function hasWindowLocalStorage(): boolean {
-  try {
-    return typeof window !== 'undefined' && !!window.localStorage;
-  } catch {
-    // Chrome throws an error if local storage is disabled and you try to access it
-    return false;
-  }
-}
