+++ src/environment.ts
@@ -1,0 +1,12 @@
+export function hasChromeStorage(): boolean {
+  return typeof chrome !== 'undefined' && !!chrome.storage;
+}
+
+export function hasWindowLocalStorage(): boolean {
+  try {
+    return typeof window !== 'undefined' && !!window.localStorage;
+  } catch {
+    // Chrome throws an error if local storage is disabled and you try to access it
+    return false;
+  }
+}
